+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased
 -----------
 - Don't panic when updating an empty A matrix.
 - Bump minimum supported Rust version to 1.63.
+- Add `CscMatrix::from_column_iter` and `CscMatrix::from_row_iter` which constructs a sparse CscMatrix with elements filled from column-major and row-major iterators.
 
 Version 0.6.2 (January 7, 2021)
 -----------
@@ -14,7 +15,7 @@ Version 0.6.1 (September 28, 2020)
 Version 0.6.0 (September 5, 2019)
 -----------
 - Update to OSQP 0.6.0 (see [OSQP changelog][osqp-clog] for details).
-- Add `CscMatrix::from_column_iter_dense` and `CscMatrix::from_row_iter_dense` to construct a dense CscMatrix with elements filled from column-major and row-major iterators.
+- Add `CscMatrix::from_column_iter_dense` and `CscMatrix::from_row_iter_dense` which constructs a dense CscMatrix with elements filled from column-major and row-major iterators.
 - `Problem` now requires the `P` matrix to be structurally upper triangular. Two methods on `CscMatrix`, `is_structurally_upper_tri` and `into_upper_tri`, are added to assist with this requirement.
 - Add variants to the `SetupError` enum returned by `Problem::new` explaining the cause of the error.